requirements.txt
crewai
langchain
langchain-openai
langchain-community
langchain-experimental
faiss-cpu
openai
fastapi
uvicorn
tiktoken
python-dotenv
pypdf
